src/test/java/leetcode/hard124/SolutionTest.java
package leetcode.hard124;

import org.junit.Test;

import static org.junit.Assert.*;

public class SolutionTest {

    @Test
    public void maxPathSum() {
        Solution solution = new Solution();
        TreeNode node1 = new TreeNode(1);
        TreeNode node2 = new TreeNode(2);
        TreeNode node3 = new TreeNode(3);
        node1.left = node2;
        node1.right = node3;
        assertEquals(6, solution.maxPathSum(node1));

        node1 = new TreeNode(-10);
        node2 = new TreeNode(9);
        node3 = new TreeNode(20);
        TreeNode node4 = new TreeNode(15);
        TreeNode node5 = new TreeNode(7);
        node1.left = node2;
        node1.right = node3;
        node3.left = node4;
        node3.right = node5;
        assertEquals(42, solution.maxPathSum(node1));
    }
}